renamed test file name to avoid a massive merge conflict
