+++ foodalert/tests.py
@@ -2,7 +2,7 @@ from django.test import TestCase, Client
 from foodalert.test.test_notification import NotificationTest
 from foodalert.test.test_subscription import SubscriptionTest
 from foodalert.test.test_update import UpdateTest
-from foodalert.test.test_sms import TwilioTest
+from foodalert.test.test_twilio import TwilioTest
 
 # Create your tests here.
